* configure.in: Check for need to provide setenv prototype if 	setenv is in the C library.
	* forward.c: Provide setenv prototype if needed.

	* auth.c (auth_name): Cast result of sizeof() to int to match
	printf format specification.

git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13580 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/telnet/libtelnet/auth.c
@@ -510,7 +510,7 @@ auth_name(data, cnt)
 	if (cnt > sizeof(savename) - 1) {
 		if (auth_debug_mode)
 			printf(">>>%s: Name in NAME (%d) exceeds %d length\r\n",
-					Name, cnt, sizeof(savename)-1);
+					Name, cnt, (int) sizeof(savename)-1);
 		return;
 	}
 	memcpy((void *)savename, (void *)data, cnt);
